refactor: modularize schedule page and shared utilities
