+++ lib/preferences.ts
@@ -1,0 +1,3 @@
+export const LANGUAGE_STORAGE_KEY = 'schedule-language';
+export const SERIES_STORAGE_KEY = 'schedule-visible-series';
+export const PERIOD_STORAGE_KEY = 'schedule-review-period-hours';
